+++ src/r_defs.h
@@ -390,6 +390,7 @@ typedef struct vissprite_s
 
   // [Nugget]
   byte *tranmap;
+  boolean fullbright;
 } vissprite_t;
 
 //