+++ src/r_voxel.c
@@ -38,6 +38,7 @@
 #include "z_zone.h"
 
 // [Nugget]
+#include "r_bsp.h"
 #include "wi_stuff.h"
 
 static boolean voxels_found;
@@ -686,6 +687,8 @@ boolean VX_ProjectVoxel (mobj_t * thing)
 	vis->x1 = x1;
 	vis->x2 = x2;
 
+	vis->fullbright = false; // [Nugget]
+
 	// get light level...
 
 	if (vis->mobjflags & MF_SHADOW)
@@ -699,6 +702,7 @@ boolean VX_ProjectVoxel (mobj_t * thing)
 	else if (thing->frame & FF_FULLBRIGHT)
 	{
 		vis->colormap[0] = vis->colormap[1] = fullcolormap;
+		vis->fullbright = true; // [Nugget]
 	}
 	else
 	{
@@ -837,6 +841,48 @@ static void VX_DrawColumn (vissprite_t * spr, int x, int y)
 	int linesize = video.pitch;
 	byte * dest = I_VideoBuffer + viewwindowy * linesize + viewwindowx;
 
+  // [Nugget] Per-column thing lighting /-------------------------------------
+
+  const byte *colormap[2];
+  memcpy(colormap, spr->colormap, sizeof(spr->colormap));
+
+  if (STRICTMODE(thing_column_lighting)
+      && !spr->fullbright && spr->colormap[0] && !fixedcolormap)
+  {
+    const fixed_t xofs = ((x << FRACBITS) + FRACUNIT/2) - v->x_pivot,
+                  yofs = ((y << FRACBITS) + FRACUNIT/2) - v->y_pivot;
+
+    const angle_t angle = (vv->angle + ANG90) >> ANGLETOFINESHIFT;
+
+    const fixed_t cosine = finecosine[angle],
+                    sine =   finesine[angle];
+
+    const fixed_t gx = spr->gx + FixedMul(xofs, cosine) + FixedMul(yofs,   sine),
+                  gy = spr->gy + FixedMul(xofs,   sine) - FixedMul(yofs, cosine);
+
+    sector_t *const sector = R_PointInSubsector(gx, gy)->sector;
+
+    int lightnum = extralight;
+
+    if (demo_version > DV_BOOM)
+    {
+      sector_t tempsector;
+      int floorlightlevel, ceilinglightlevel;
+
+      R_FakeFlat(sector, &tempsector, &floorlightlevel, &ceilinglightlevel, false);
+
+      lightnum += (floorlightlevel + ceilinglightlevel) >> (LIGHTSEGSHIFT+1);
+    }
+    else { lightnum += sector->lightlevel >> LIGHTSEGSHIFT; }
+
+    const int lightindex = STRICTMODE(!diminished_lighting)
+                           ? 0 : R_GetLightIndex(B_xscale);
+
+    colormap[0] = scalelight[BETWEEN(0, LIGHTLEVELS-1, lightnum)][lightindex];
+  }
+
+  // [Nugget] ---------------------------------------------------------------/
+
 	// iterate over screen columns
 	fixed_t ux = ((Ax - 1) | FRACMASK) + 1;
 
@@ -927,7 +973,7 @@ static void VX_DrawColumn (vissprite_t * spr, int x, int y)
 					uy = clip_y1;
 
 				byte src = slab[0];
-				byte pix = spr->colormap[spr->brightmap[src]][src];
+				byte pix = colormap[spr->brightmap[src]][src];
 
 				for (; uy < uy1 ; uy += FRACUNIT)
 				{
@@ -942,7 +988,7 @@ static void VX_DrawColumn (vissprite_t * spr, int x, int y)
 					uy = clip_y2;
 
 				byte src = slab[len - 1];
-				byte pix = spr->colormap[spr->brightmap[src]][src];
+				byte pix = colormap[spr->brightmap[src]][src];
 
 				for (; uy > uy2 ; uy -= FRACUNIT)
 				{
@@ -962,7 +1008,7 @@ static void VX_DrawColumn (vissprite_t * spr, int x, int y)
 					if (i >= len) i = len - 1;
 
 					byte src = slab[i];
-					byte pix = spr->colormap[spr->brightmap[src]][src];
+					byte pix = colormap[spr->brightmap[src]][src];
 
 					dest[(uy >> FRACBITS) * linesize + (ux >> FRACBITS)] = pix;
 				}